+++ src/main/java/lotto/view/input/UserInput.java
@@ -5,22 +5,22 @@ import lotto.validation.Validator;
 
 public class UserInput {
 
-    public long getLottoCount(){
-        long price = getPrice();
+    public int getLottoCount(){
+        int price = getPrice();
         return price / 1000;
     }
 
-    private Long getPrice() {
-        long price = getNumericPrice();
+    private int getPrice() {
+        int price = getNumericPrice();
         Validator.validatePriceRange(price);
         Validator.validateUnit(price);
         return price;
     }
 
-    private Long getNumericPrice() {
+    private int getNumericPrice() {
         String userInput = getUserInput();
         Validator.validateStringIsNumeric(userInput);
-        return Long.parseLong(userInput);
+        return Integer.parseInt(userInput);
     }
 
     private String getUserInput() {
